Former-commit-id: 702119dad4539e847dfedad35ccb6cb45f9ff0a4

+++ src/gamestatus/WriteableGameStatus.java
@@ -2,8 +2,8 @@ package gamestatus;
 
 public interface WriteableGameStatus {
 	void addExperience(double exp);
-	void addMoney(double mon);
-	void spendMoney(double mon);
+	void addMoney(int mon);
+	void spendMoney(int mon);
 	void loseLife();
 	void gainLife();
 	void setMyEnemiesLeft(int enemiesLeft);
